requirements.txt
dnspython>=2.6.1
python-whois>=0.9.5
requests>=2.32.3
tldextract>=5.1.2
beautifulsoup4>=4.12.3
builtwith>=1.3.4
reportlab>=4.2.2
fastapi>=0.115.0
uvicorn>=0.30.6
jinja2>=3.1.4
python-multipart>=0.0.9
